Creating a CreditCardReposity to contain ElasticSearch
